Rename Fugu.Storage -> Fugu.IO
