+++ src/Fugu.Core/IO/IBackingStorage.cs
@@ -1,4 +1,4 @@
-﻿namespace Fugu.Storage;
+﻿namespace Fugu.IO;
 
 public interface IBackingStorage
 {
