+++ src/Fugu.Core/IO/ISlab.cs
@@ -1,4 +1,4 @@
-﻿namespace Fugu.Storage;
+﻿namespace Fugu.IO;
 
 /*
  * Names:
