+++ src/Fugu.Core/IO/IWritableSlab.cs
@@ -1,4 +1,4 @@
-﻿namespace Fugu.Storage;
+﻿namespace Fugu.IO;
 
 public interface IWritableSlab : ISlab
 {
